Initiate Project
The pieces of code I used in a previous class, for Databases II.
It's a website based in Javascript, featuring a search engine for songs,
connecting the user to the appropriate video.
Also there is a control panel for a musician to add his video, with the
description and the musicians of the other instruments on the song, so
that a user can search for it.
It is built on NetBeans 7.3 with Glassfish's support.