Most work completed.No checkout,triggers,car/room
Work completed:
-Index (Dropdown boxes filled)
-Results (According to database)
-UserCP (change data, cancel flights according to database)
-Login/Verify Process (PopUp Box/Account on database)
-Cookie Work (Login username cookie, remember me)
-Checkout Security (Logged-in Check)
Left to do:
-Complete checkout
-Change/add/remove something
-Success
-Minor general improvements
-Everything involving rooms, cars and bargains.
Now, you can attemp to stlyle all of the pages, but if a change is to be
done on the coding I want to know first. Change of extensions can be
made freely, if and only if the outcome is the same operationally.
Questions will be asked online!